App/backend-v2/extract-transform-load/requirements.txt
apache-airflow==2.9.2
kubernetes==30.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-core==2.16.3